+++ test/_support/targets.ex
@@ -5,10 +5,9 @@ defmodule TypeTest.Targets do
   alias Type.{Bitstring, Function, List, Map, Tuple}
 
   @example_list [-47, builtin(:neg_integer), 0, 47, -10..10,
-  builtin(:pos_integer), builtin(:non_neg_integer), builtin(:integer),
-  builtin(:float), :foo, builtin(:atom), builtin(:reference),
-  %Function{params: [], return: 0}, builtin(:port), builtin(:pid),
-  %Tuple{elements: []}, Map.build(%{builtin(:any) => builtin(:any)}),
+  builtin(:pos_integer), builtin(:float), :foo, builtin(:atom),
+  builtin(:reference), %Function{params: [], return: 0}, builtin(:port),
+  builtin(:pid), %Tuple{elements: []}, Map.build(%{builtin(:any) => builtin(:any)}),
   [], %List{}, %Bitstring{size: 0, unit: 0}]
 
   def except(list \\ []) do