+++ test/type/type_opaque/usable_as_test.exs
@@ -22,8 +22,8 @@ defmodule TypeTest.TypeOpaque.UsableAsTest do
     end
 
     test "warns when usable_as is transferred" do
-      assert {:warn, _} = @opaque_int ~> builtin(:integer)
-      assert {:warn, _} = builtin(:integer) ~> @opaque_int
+      assert {:maybe, _} = @opaque_int ~> builtin(:integer)
+      assert {:maybe, _} = builtin(:integer) ~> @opaque_int
     end
   end
 end